pyproject.toml
[project]
name = "REMIND-PyPSA-coupling"
version = "0.0.0a3"
description = "Tools to couple the REMIND IAM and PyPSA power models"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>1.9", 
    "pandas>2", 
    "pyyaml", 
    "country_converter"]

[project.urls]
Homepage = "https://github.com/pik-piam/Remind-PyPSA-coupling/"
Documentation = "https://pik-piam.github.io/Remind-PyPSA-coupling/"

# eg `uv pip install -r pyproject.toml --extra release`
[project.optional-dependencies]
jupyter = ["ipykernel","ipython"]
gdx = ["gamspy"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = [    
    "setuptools", "wheel", "numpy", "cython", "twine"
]
build-backend = "setuptools.build_meta"




